refactor: move about package to internal/about
