+++ internal/about/util.go
