+++ internal/gotestast/spec.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/dlclark/regexp2"
 
-	"github.com/mvrahden/go-test/about"
+	"github.com/mvrahden/go-test/internal/about"
 	"github.com/mvrahden/go-test/internal/x/slices"
 	"golang.org/x/tools/go/packages"
 )
