+++ internal/gotestgen/sharedfixture.go
@@ -6,7 +6,7 @@ import (
 	"go/format"
 	"text/template"
 
-	"github.com/mvrahden/go-test/about"
+	"github.com/mvrahden/go-test/internal/about"
 )
 
 var sharedFixtureTpl = template.Must(template.New("sharedfixture").ParseFS(templates, "static/sharedfixture.go.tpl"))
