+++ internal/gotestrunner/gotestrunner_suite_test.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/mvrahden/go-test/about"
+	"github.com/mvrahden/go-test/internal/about"
 	"github.com/mvrahden/go-test/internal/gotestgen"
 	"github.com/mvrahden/go-test/internal/gotestrunner"
 	"github.com/mvrahden/go-test/internal/protocol"
